requirements.txt
python-nmap
gspread
oauth2client